requirements.txt
pytz
pandas
requests
beautifulsoup4
pyyaml
setuptools
